+++ lofty_attr/src/lib.rs
@@ -15,6 +15,7 @@ const LOFTY_FILE_TYPES: [&str; 10] = [
 /// Creates a file usable by Lofty
 ///
 /// See [here](https://github.com/Serial-ATA/lofty-rs/tree/main/examples/custom_resolver) for an example of how to use it.
+// TODO: #[internal]
 #[proc_macro_derive(LoftyFile, attributes(lofty))]
 pub fn lofty_file(input: TokenStream) -> TokenStream {
 	let input = parse_macro_input!(input as DeriveInput);